\name{plot.predcurve}
\alias{plot.predcurve}
\title{Plots the predictiveness curve}
\usage{
  plot.predcurve(R, vl = NULL, vh = NULL, ...)
}
\arguments{
  \item{R}{An object of class "predcurve"}

  \item{...}{other arguments passed to plot}
}
\value{
  output NULL
}
\description{
  Plots the predictiveness curve
}
\keyword{accuracy}
\keyword{curve}
\keyword{prediction}
\keyword{predictiveness}

