\docType{package}
\name{predcurver}
\alias{predcurver}
\alias{predcurver-package}
\title{Class and methods for the predictiveness curve}
\description{
  Estimates the predictiveness curve from a vector of risks
  (true or estimated). Plots, prints, summarizes, and tests
  using the predcurve S3 class.
}

