\name{print.predcurve}
\alias{print.predcurve}
\title{Prints the predictiveness curve}
\usage{
  print.predcurve(R, ...)
}
\arguments{
  \item{R}{An object of class "predcurve"}
}
\value{
  output NULL
}
\description{
  Prints the predictiveness curve
}
\keyword{accuracy}
\keyword{curve}
\keyword{prediction}
\keyword{predictiveness}

